triplifier/requirements.txt
certifi==2025.4.26
chardet==5.2.0
Flask==3.1.1
Flask-Compress==1.17
Flask-SQLAlchemy==3.1.1
gevent==25.5.1
gunicorn==23.0.0
numpy==2.2.6
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
# Python Triplifier dependencies
PyYAML>=5.3
rdflib>=6.0.0
SQLAlchemy>=1.4